+++ BankAccountDDD/Account.DDD/Commands/DeposeCheque.cs
@@ -8,8 +8,8 @@ namespace Account.DDD.Commands
     public class DeposeCheque : Command
     {
         public Guid AccountId;
-        public double Funds;
-        public DeposeCheque(Guid id, double funds)
+        public decimal Funds;
+        public DeposeCheque(Guid id, decimal funds)
         {
             AccountId = id;
             Funds = funds;
